Make sure that the writing task queue is not unbounded
This commit adds a bounded queue/pool which via delegation strictly
limits the number of in flight tasks.

In most cases on modern hardware with high frequency and multi-core CPUs
it will be faster to read and decompress data from the source than to
write it out to its destination.  This is true regardless of destination
file format or the presence of high performance storage.

A `ThreadPoolExecutor` in its standard form maintains an unbounded queue
for tasks.  Thus, `ThreadPoolExecutor.submit()` will never block and it
will happily take new tasks until system memory is exhausted.  That's
**very** bad for our use case as it rather quickly results in a very
large number tasks operating on 1.5MiB or more of data being
accumulated.  GiBs of system memory is then consumed.