Fix some logic errors
Firstly, `NamedTemporaryFile` cannot be opened more than once on
Windows; the implementation was effectively broken completely on the
platform.  Consequently, temporary file usage has been moved to
`mkstemp()` and handled accordingly.  Furthermore, the `out` argument to
the JPEG-2000 codec `Codec.decode()` implementation was not being
handled correctly causing visual corruption with certain workflows
(noteably napari).

As of this commit glymur 0.9.4 is required and the constructor based API
is now in use.